+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-## 1.25.2
+## 1.26.0
+
+* Don't throw errors if the exact same member is loaded or forwarded from
+  multiple modules at the same time.
 
 * Fix a bug where, under extremely rare circumstances, a valid variable could
   become unassigned.